Refactored getting topology.
